requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
readability-lxml==0.8.1
pydantic==2.8.2
tenacity==8.5.0
tldextract==5.1.2
psycopg[binary]==3.2.9
python-dotenv==1.0.1
